Update File Structure
